refactor(http): split stream and sse files
Split the streaming and SSE client surfaces into dedicated source files and finish the HTTP file renumbering so the module sorts in execution order.

- move Stream/open/startread/closeread into src/7_6_http_stream.jl
- move SSE parsing and SSEEvent into src/7_6_http_sse.jl
- rename HTTP sources to the 7_x pattern and update includes/docs

Reference: PR follow-up cleanup
