+++ src/7_6_http_client.jl
@@ -630,10 +630,6 @@ export MemoryCookieJar
 export Cookie
 export do!
 export get!
-export Stream
-export startread
-export closeread
-export SSEEvent
 export StatusError
 export request
 export get
@@ -643,7 +639,6 @@ export put
 export patch
 export delete
 export options
-export open
 
 abstract type AbstractCookieJar end
 
@@ -1292,7 +1287,7 @@ function get!(client::Client, address::AbstractString, target::AbstractString; s
     return do!(client, address, request; secure = secure, protocol = protocol)
 end
 
-import Base: close, closewrite, eof, get, isopen, open, read, readbytes!, write
+import Base: get
 
 """
     StatusError
@@ -1751,280 +1746,6 @@ function _apply_conn_deadline!(conn::ClientConn, deadline_ns::Int64)
     return nothing
 end
 
-const _SSE_LF = UInt8('\n')
-const _SSE_CR = UInt8('\r')
-const _SSE_COLON = UInt8(':')
-const _SSE_SPACE = UInt8(' ')
-
-"""
-    SSEEvent
-
-Parsed Server-Sent Event delivered to `sse_callback`.
-
-Fields:
-- `data`: concatenated `data:` lines joined by `\n`
-- `event`: optional event name
-- `id`: last event id in effect when the event dispatched
-- `retry`: last valid retry hint in milliseconds
-- `fields`: all observed fields collected into a string dictionary
-"""
-struct SSEEvent
-    data::String
-    event::Union{Nothing, String}
-    id::Union{Nothing, String}
-    retry::Union{Nothing, Int}
-    fields::Dict{String, String}
-end
-
-mutable struct _SSEState
-    data_lines::Vector{String}
-    has_data::Bool
-    bom_checked::Bool
-    saw_evidence::Bool
-    event_name::Union{Nothing, String}
-    last_event_id::Union{Nothing, String}
-    last_retry::Union{Nothing, Int}
-    fields::Dict{String, Vector{String}}
-end
-
-function _SSEState()
-    return _SSEState(String[], false, false, false, nothing, nothing, nothing, Dict{String, Vector{String}}())
-end
-
-function _ensure_sse_content_type(headers::Headers)::Nothing
-    ctype = get_header(headers, "Content-Type")
-    ctype === nothing && throw(ErrorException("Response Content-Type is not text/event-stream"))
-    base = strip(first(split(ctype, ';')))
-    lowercase(base) == "text/event-stream" || throw(ErrorException("Response Content-Type is not text/event-stream"))
-    return nothing
-end
-
-function _wrap_sse_callback(callback::Function, response::Response)::Function
-    dummy = SSEEvent("", nothing, nothing, nothing, Dict{String, String}())
-    if applicable(callback, response, dummy)
-        return event -> callback(response, event)
-    end
-    applicable(callback, dummy) || throw(ArgumentError("sse_callback must accept (event) or (response, event)"))
-    return callback
-end
-
-function _trim_sse_line_length(bytes::AbstractVector{UInt8})::Int
-    len = length(bytes)
-    while len > 0 && bytes[len] == _SSE_CR
-        len -= 1
-    end
-    return len
-end
-
-function _sse_bytes_to_string(bytes::AbstractVector{UInt8})::String
-    isempty(bytes) && return ""
-    try
-        return String(bytes)
-    catch err
-        err isa ArgumentError && throw(ErrorException("SSE stream emitted invalid UTF-8 data"))
-        rethrow(err)
-    end
-end
-
-function _looks_like_sse_prefix(partial::Vector{UInt8})::Bool
-    isempty(partial) && return false
-    idx = 1
-    if length(partial) >= 3 && partial[1] == 0xEF && partial[2] == 0xBB && partial[3] == 0xBF
-        idx = 4
-    end
-    idx > length(partial) && return false
-    b = partial[idx]
-    b == _SSE_COLON && return true
-    remaining = length(partial) - idx + 1
-    remaining >= 4 && partial[idx] == UInt8('d') && partial[idx + 1] == UInt8('a') &&
-        partial[idx + 2] == UInt8('t') && partial[idx + 3] == UInt8('a') && return true
-    remaining >= 5 && partial[idx] == UInt8('e') && partial[idx + 1] == UInt8('v') &&
-        partial[idx + 2] == UInt8('e') && partial[idx + 3] == UInt8('n') && partial[idx + 4] == UInt8('t') &&
-        return true
-    remaining >= 2 && partial[idx] == UInt8('i') && partial[idx + 1] == UInt8('d') && return true
-    remaining >= 5 && partial[idx] == UInt8('r') && partial[idx + 1] == UInt8('e') &&
-        partial[idx + 2] == UInt8('t') && partial[idx + 3] == UInt8('r') && partial[idx + 4] == UInt8('y') &&
-        return true
-    return false
-end
-
-function _append_sse_field!(state::_SSEState, field::String, value::String)::Nothing
-    values = get(() -> String[], state.fields, field)
-    push!(values, value)
-    state.fields[field] = values
-    return nothing
-end
-
-function _parse_sse_retry(value::String)
-    parsed = tryparse(Int, strip(value))
-    (parsed === nothing || parsed < 0) && return nothing
-    return parsed
-end
-
-function _process_sse_field!(state::_SSEState, field::String, value::String)::Nothing
-    if field == "data"
-        state.saw_evidence = true
-        _append_sse_field!(state, field, value)
-        push!(state.data_lines, value)
-        state.has_data = true
-        return nothing
-    end
-    if field == "event"
-        state.saw_evidence = true
-        _append_sse_field!(state, field, value)
-        state.event_name = value
-        return nothing
-    end
-    if field == "id"
-        state.saw_evidence = true
-        occursin('\0', value) && return nothing
-        _append_sse_field!(state, field, value)
-        state.last_event_id = value
-        return nothing
-    end
-    if field == "retry"
-        state.saw_evidence = true
-        _append_sse_field!(state, field, value)
-        parsed = _parse_sse_retry(value)
-        parsed === nothing || (state.last_retry = parsed)
-        return nothing
-    end
-    _append_sse_field!(state, field, value)
-    return nothing
-end
-
-function _build_sse_fields(state::_SSEState)::Dict{String, String}
-    out = Dict{String, String}()
-    for (key, entries) in state.fields
-        isempty(entries) && continue
-        out[key] = length(entries) == 1 ? entries[1] : join(entries, "\n")
-    end
-    return out
-end
-
-function _reset_sse_event!(state::_SSEState)::Nothing
-    empty!(state.data_lines)
-    state.has_data = false
-    state.event_name = nothing
-    for values in values(state.fields)
-        empty!(values)
-    end
-    return nothing
-end
-
-function _emit_sse_event!(state::_SSEState, callback::Function)::Nothing
-    data = isempty(state.data_lines) ? "" : join(state.data_lines, "\n")
-    event = SSEEvent(data, state.event_name, state.last_event_id, state.last_retry, _build_sse_fields(state))
-    callback(event)
-    return nothing
-end
-
-function _dispatch_sse_event!(state::_SSEState, callback::Function)::Nothing
-    if state.has_data
-        _emit_sse_event!(state, callback)
-    end
-    _reset_sse_event!(state)
-    return nothing
-end
-
-function _process_sse_line!(state::_SSEState, raw::AbstractVector{UInt8}, callback::Function)::Nothing
-    line_len = _trim_sse_line_length(raw)
-    if line_len == 0
-        _dispatch_sse_event!(state, callback)
-        return nothing
-    end
-    line = line_len == length(raw) ? raw : @view raw[1:line_len]
-    if !state.bom_checked
-        state.bom_checked = true
-        if length(line) >= 3 && line[1] == 0xEF && line[2] == 0xBB && line[3] == 0xBF
-            line = length(line) == 3 ? UInt8[] : @view line[4:length(line)]
-            isempty(line) && return nothing
-        end
-    end
-    line[1] == _SSE_COLON && (state.saw_evidence = true; return nothing)
-    colon_idx = findfirst(isequal(_SSE_COLON), line)
-    if colon_idx === nothing
-        field = _sse_bytes_to_string(line)
-        value = ""
-    else
-        field = colon_idx == 1 ? "" : _sse_bytes_to_string(@view line[1:colon_idx - 1])
-        value_slice = colon_idx == length(line) ? UInt8[] : @view line[colon_idx + 1:length(line)]
-        if !isempty(value_slice) && value_slice[1] == _SSE_SPACE
-            value_slice = length(value_slice) == 1 ? UInt8[] : @view value_slice[2:length(value_slice)]
-        end
-        value = _sse_bytes_to_string(value_slice)
-    end
-    _process_sse_field!(state, field, value)
-    return nothing
-end
-
-function _process_sse_chunk!(state::_SSEState, partial::Vector{UInt8}, chunk::AbstractVector{UInt8}, callback::Function)::Nothing
-    start = 1
-    len = length(chunk)
-    while start <= len
-        nl = findnext(isequal(_SSE_LF), chunk, start)
-        nl === nothing && break
-        stop = nl - 1
-        if isempty(partial)
-            if stop >= start
-                _process_sse_line!(state, @view(chunk[start:stop]), callback)
-            else
-                _process_sse_line!(state, UInt8[], callback)
-            end
-        else
-            if stop >= start
-                append!(partial, @view(chunk[start:stop]))
-            end
-            _process_sse_line!(state, partial, callback)
-            empty!(partial)
-        end
-        start = nl + 1
-    end
-    start <= len && append!(partial, @view(chunk[start:len]))
-    return nothing
-end
-
-function _parse_sse_stream!(io::IO, callback::Function)::Int64
-    state = _SSEState()
-    partial = UInt8[]
-    total = Int64(0)
-    detect_bytes = 8 * 1024
-    while !eof(io)
-        chunk = readavailable(io)
-        isempty(chunk) && continue
-        total += length(chunk)
-        _process_sse_chunk!(state, partial, chunk, callback)
-        if total >= detect_bytes && !state.saw_evidence && !_looks_like_sse_prefix(partial)
-            throw(ErrorException("Response does not appear to be a Server-Sent Events stream"))
-        end
-    end
-    if !isempty(partial)
-        _process_sse_line!(state, partial, callback)
-        empty!(partial)
-    end
-    total > 0 && !state.saw_evidence && throw(ErrorException("Response does not appear to be a Server-Sent Events stream"))
-    if state.has_data
-        _emit_sse_event!(state, callback)
-        _reset_sse_event!(state)
-    end
-    return total
-end
-
-function _consume_incoming_sse!(
-        incoming::_IncomingResponse,
-        response::Response,
-        callback::Function;
-        decompress::Union{Nothing, Bool},
-    )::Nothing
-    _ensure_sse_content_type(incoming.head.headers)
-    wrapped = _wrap_sse_callback(callback, response)
-    _with_response_reader(incoming; decompress = decompress) do reader
-        _parse_sse_stream!(reader, wrapped)
-        return nothing
-    end
-    return nothing
-end
-
 """
     request(method, url, headers=Pair{String,String}[], body=nothing; kwargs...)
 
@@ -2157,329 +1878,6 @@ function _finalize_request_response(
         request = resolved_request,
         request_url = request_url,
     )
-end
-
-"""
-    Stream <: IO
-
-Client-side request/response stream returned by `HTTP.open`.
-
-Writes append request body bytes until response reading begins. After
-`startread(stream)`, reads consume the response body from the underlying
-connection using the same redirect/decompression machinery as `request(...)`.
-"""
-mutable struct Stream <: IO
-    method::String
-    parsed::_URLParts
-    headers::Headers
-    client::Client
-    owns_client::Bool
-    redirect::Bool
-    status_exception::Bool
-    protocol::Symbol
-    decompress::Union{Nothing, Bool}
-    readtimeout::Float64
-    request_buffer::IOBuffer
-    response::Union{Nothing, Response}
-    reader::Union{Nothing, IO}
-    producer::Union{Nothing, Task}
-    @atomic started::Bool
-    @atomic write_closed::Bool
-    @atomic read_closed::Bool
-end
-
-function Stream(
-        method::AbstractString,
-        parsed::_URLParts,
-        headers::Headers,
-        client::Client,
-        owns_client::Bool;
-        redirect::Bool,
-        status_exception::Bool,
-        protocol::Symbol,
-        decompress::Union{Nothing, Bool},
-        readtimeout::Real,
-    )
-    readtimeout >= 0 || throw(ArgumentError("readtimeout must be >= 0"))
-    return Stream(
-        String(method),
-        parsed,
-        headers,
-        client,
-        owns_client,
-        redirect,
-        status_exception,
-        protocol,
-        decompress,
-        Float64(readtimeout),
-        IOBuffer(),
-        nothing,
-        nothing,
-        nothing,
-        false,
-        false,
-        false,
-    )
-end
-
-function _stream_response(stream::Stream)::Response
-    resp = stream.response
-    resp === nothing && throw(ProtocolError("response has not started yet"))
-    return resp::Response
-end
-
-function _stream_reader(stream::Stream)::IO
-    reader = stream.reader
-    reader === nothing && throw(ProtocolError("response body reader is not available"))
-    return reader::IO
-end
-
-function _finish_stream_read!(stream::Stream; suppress_producer_errors::Bool)::Response
-    was_closed = @atomic :acquire stream.read_closed
-    was_closed && return _stream_response(stream)
-    @atomic :release stream.read_closed = true
-    reader = stream.reader
-    producer = stream.producer
-    try
-        if reader !== nothing
-            close(reader)
-        end
-    catch
-    end
-    if producer !== nothing
-        if suppress_producer_errors
-            try
-                wait(producer)
-            catch
-            end
-        else
-            wait(producer)
-        end
-    end
-    if stream.owns_client
-        close(stream.client)
-    end
-    return _stream_response(stream)
-end
-
-function _start_stream_read!(stream::Stream)::Response
-    started = @atomic :acquire stream.started
-    started && return _stream_response(stream)
-    @atomic :release stream.started = true
-    @atomic :release stream.write_closed = true
-    request_bytes = take!(stream.request_buffer)
-    body_input = isempty(request_bytes) ? nothing : request_bytes
-    req_body, content_length = _normalize_body_input(body_input)
-    req = Request(
-        stream.method,
-        stream.parsed.target;
-        headers = stream.headers,
-        body = req_body,
-        host = stream.parsed.address,
-        content_length = content_length,
-    )
-    if stream.readtimeout > 0
-        timeout_ns = Int64(round(stream.readtimeout * 1.0e9))
-        set_deadline!(req.context, Int64(time_ns()) + timeout_ns)
-    end
-    incoming = if stream.redirect
-        _do_incoming!(
-            stream.client,
-            stream.parsed.address,
-            req;
-            secure = stream.parsed.secure,
-            server_name = stream.parsed.server_name,
-            protocol = stream.protocol,
-        )
-    else
-        _roundtrip_incoming!(
-            stream.client.transport,
-            stream.parsed.address,
-            req;
-            secure = stream.parsed.secure,
-            server_name = stream.parsed.server_name,
-        )
-    end
-    resolved_request = incoming.head.request === nothing ? req : incoming.head.request::Request
-    stream.response = _finalize_request_response(
-        incoming,
-        nothing,
-        _should_decompress_response(incoming.head.headers, stream.decompress) ? Int64(-1) : incoming.head.content_length,
-        resolved_request,
-        stream.parsed.url,
-    )
-    reader, producer = _response_body_reader(incoming; decompress = stream.decompress)
-    stream.reader = reader
-    stream.producer = producer
-    return stream.response::Response
-end
-
-"""
-    startread(stream) -> Response
-
-Finalize request writes if needed, execute the HTTP exchange, and return the
-response metadata for `stream` without buffering the response body.
-
-Subsequent reads on `stream` consume the response body. Repeated calls return
-the same response object.
-"""
-function startread(stream::Stream)::Response
-    return _start_stream_read!(stream)
-end
-
-function isopen(stream::Stream)::Bool
-    return !(@atomic :acquire stream.read_closed) || !(@atomic :acquire stream.write_closed)
-end
-
-function write(stream::Stream, data::AbstractVector{UInt8})::Int
-    (@atomic :acquire stream.started) && throw(ArgumentError("cannot write request body after response reading has started"))
-    (@atomic :acquire stream.write_closed) && throw(ArgumentError("request body writes are closed"))
-    return write(stream.request_buffer, data)
-end
-
-function write(stream::Stream, data::Union{String, SubString{String}})::Int
-    (@atomic :acquire stream.started) && throw(ArgumentError("cannot write request body after response reading has started"))
-    (@atomic :acquire stream.write_closed) && throw(ArgumentError("request body writes are closed"))
-    return write(stream.request_buffer, data)
-end
-
-function closewrite(stream::Stream)
-    @atomic :release stream.write_closed = true
-    return nothing
-end
-
-function readbytes!(stream::Stream, dest::AbstractVector{UInt8}, nb::Integer = length(dest))
-    nb >= 0 || throw(ArgumentError("nb must be >= 0"))
-    _start_stream_read!(stream)
-    n = readbytes!(_stream_reader(stream), dest, nb)
-    n == 0 && _finish_stream_read!(stream; suppress_producer_errors = false)
-    return n
-end
-
-function read(stream::Stream)::Vector{UInt8}
-    _start_stream_read!(stream)
-    bytes = read(_stream_reader(stream))
-    _finish_stream_read!(stream; suppress_producer_errors = false)
-    return bytes
-end
-
-function read(stream::Stream, ::Type{String})::String
-    return String(read(stream))
-end
-
-function eof(stream::Stream)::Bool
-    _start_stream_read!(stream)
-    done = eof(_stream_reader(stream))
-    done && _finish_stream_read!(stream; suppress_producer_errors = false)
-    return done
-end
-
-"""
-    closeread(stream) -> Response
-
-Close the readable side of `stream` and return its response metadata.
-
-If the response body has already been fully consumed, this is effectively a
-no-op. If unread response bytes remain, the underlying client connection is not
-reused.
-"""
-function closeread(stream::Stream)::Response
-    _start_stream_read!(stream)
-    return _finish_stream_read!(stream; suppress_producer_errors = true)
-end
-
-function close(stream::Stream)
-    try
-        closewrite(stream)
-    catch
-    end
-    try
-        closeread(stream)
-    catch
-    end
-    return nothing
-end
-
-"""
-    open(method::Symbol, url, headers=Pair{String,String}[]; kwargs...) -> Stream
-    open(f, method::Symbol, url, headers=Pair{String,String}[]; kwargs...)
-
-Create a streaming HTTP client request/response exchange.
-
-The returned `Stream` buffers request writes locally until `startread(stream)`
-or the end of the `do` block. Once reading starts, `stream` behaves like a
-readable `IO` for the response body. `kwargs` largely mirror `request(...)`,
-including `redirect`, `decompress`, `client`, `connect_timeout`, `readtimeout`,
-`require_ssl_verification`, and `protocol`.
-
-The `do`-block form closes request writes automatically, closes the readable
-side on exit, and returns the final response metadata.
-
-Method is currently a `Symbol` to avoid colliding with Base's file-opening
-`open(::AbstractString, ::AbstractString)` methods during precompilation.
-"""
-function open(
-        method::Symbol,
-        url::AbstractString,
-        headers = Pair{String, String}[];
-        status_exception::Bool = true,
-        redirect::Bool = true,
-        query = nothing,
-        decompress::Union{Nothing, Bool} = nothing,
-        client::Union{Nothing, Client} = nothing,
-        connect_timeout::Real = 0,
-        readtimeout::Real = 0,
-        require_ssl_verification::Bool = true,
-        protocol::Symbol = :auto,
-        kwargs...,
-    )::Stream
-    _validate_request_extra_kwargs(kwargs)
-    parsed = _parse_http_url(url; query = query)
-    req_headers = _normalize_headers_input(headers)
-    _apply_default_accept_encoding!(req_headers, decompress)
-    if parsed.authorization !== nothing && !has_header(req_headers, "Authorization")
-        set_header!(req_headers, "Authorization", parsed.authorization::String)
-    end
-    req_client, owns_client = _client_for_request(client; connect_timeout = connect_timeout, require_ssl_verification = require_ssl_verification)
-    return Stream(
-        _method_upper(String(method)),
-        parsed,
-        req_headers,
-        req_client,
-        owns_client;
-        redirect = redirect,
-        status_exception = status_exception,
-        protocol = protocol,
-        decompress = decompress,
-        readtimeout = readtimeout,
-    )
-end
-
-function open(
-        f::Function,
-        method::Symbol,
-        url::AbstractString,
-        headers = Pair{String, String}[];
-        kwargs...,
-    )
-    stream = open(method, url, headers; kwargs...)
-    callback_error = nothing
-    try
-        f(stream)
-    catch err
-        callback_error = err
-    finally
-        try
-            closewrite(stream)
-        catch
-        end
-    end
-    response = closeread(stream)
-    if stream.status_exception && _status_throws(response)
-        throw(StatusError(response))
-    end
-    callback_error === nothing || throw(callback_error)
-    return response
 end
 
 function _split_headers_body_args(args::Tuple)
